Commands and refactoring - Added !!/hug and !!/cookie - Moved typeclasses and command regexes out of util.py and into their own files - Formatted a few things with black
